# apifox-openapi.yaml
openapi: 3.0.3
info:
  title: 剧本杀拼车系统 API
  version: 2.4.0
  description: |
    JuPin 后端接口设计文档，可直接导入 Apifox。

    本地测试 Base URL: http://localhost:8080

    前端和 Nginx 未启动不影响接口测试，Apifox 直接请求 Spring Boot 后端即可。

    认证方式：
    - 注册/登录接口不需要 Token
    - 玩家端接口使用玩家 Token
    - 店家端接口使用店家 Token
    - 管理端接口使用管理员 Token
    - Header: Authorization: Bearer <accessToken>

    WebSocket:
    - 连接端点：/ws
    - 拼车状态订阅：/topic/pool/{poolId}
    - 群聊订阅：/topic/pool/{poolId}/chat
    - 匹配推送：/queue/match/{userId}
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              phone: '13800000001'
              password: abc123456
              nickname: 玩家一号
              gender: 1
              role: player
              city: 上海
      responses:
        '200':
          description: 注册成功，返回 Access Token、Refresh Token 和用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
  /api/auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              phone: '13800000001'
              password: abc123456
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
...
  /api/player/script/list:
    get:
      tags: [玩家端-剧本]
      summary: 系统剧本库列表
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 剧本列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptListResult'
...
  /api/player/order/create:
    post:
      tags: [玩家端-订单]
      summary: 创建订单
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
            example:
              poolId: 1
              type: 0
      responses:
        '200':
          description: 订单信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResult'
  /api/player/order/pay/{orderNo}:
    post:
      tags: [玩家端-订单]
      summary: 模拟支付
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
          example: '202605210001'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/order/my:
    get:
      tags: [玩家端-订单]
      summary: 我的订单
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: integer
          description: 0-押金 1-车费
        - name: status
          in: query
          required: false
          schema:
            type: integer
          description: 0-待支付 1-已支付 2-已退款 3-已扣留 4-逾期
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 我的订单列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResult'
...
  /api/shop/script/list:
    get:
      tags: [店家端-剧本]
      summary: 系统剧本库
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 剧本列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptListResult'
...
  /api/shop/script/{shopId}/scripts/{scriptId}:
    delete:
      tags: [店家端-剧本]
      summary: 移除店铺剧本
      parameters:
        - $ref: '#/components/parameters/ShopId'
        - $ref: '#/components/parameters/ScriptId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/admin/script/create:
    post:
      tags: [管理端-剧本]
      summary: 创建剧本
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptCreateRequest'
            example:
              name: 年轮
              type: 硬核
              difficulty: 2
              minPlayers: 4
              maxPlayers: 6
              duration: 240
              roles: '[{"name":"侦探"}]'
              cover: ''
              priceRef: 88
              description: 硬核推理本
      responses:
        '200':
          description: 剧本信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptResult'
  /api/admin/script/{scriptId}:
    put:
      tags: [管理端-剧本]
      summary: 修改剧本
      parameters:
        - $ref: '#/components/parameters/ScriptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptCreateRequest'
      responses:
        '200':
          description: 剧本信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptResult'
    delete:
      tags: [管理端-剧本]
      summary: 下架/删除剧本
      parameters:
        - $ref: '#/components/parameters/ScriptId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/admin/script/list:
    get:
      tags: [管理端-剧本]
      summary: 剧本列表
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 剧本列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptListResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ShopId:
      name: shopId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    ScriptId:
      name: scriptId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
      example: 1
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 10
      example: 10
...
  responses:
    Ok:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResult'
  schemas:
    BaseResult:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
    LoginResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoginVO'
...
    OrderResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderVO'
    OrderListResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/OrderVO'
    ScriptResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ScriptVO'
    ScriptListResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ScriptVO'
...
    RegisterRequest:
      type: object
      required: [phone, password, nickname, role]
      properties:
        phone:
          type: string
          example: '13800000001'
        password:
          type: string
          example: abc123456
        nickname:
          type: string
          example: 玩家一号
        gender:
          type: integer
          example: 1
        role:
          type: string
          enum: [player, shop]
          example: player
        city:
          type: string
          example: 上海
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
          example: '13800000001'
        password:
          type: string
          example: abc123456
    LoginVO:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserVO'
    UserVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        phone:
          type: string
        nickname:
          type: string
        avatar:
          type: string
        gender:
          type: integer
        role:
          type: integer
          description: 0-玩家 1-店家 2-管理员
        city:
          type: string
        preference:
          type: string
        creditScore:
          type: integer
        status:
          type: integer
...
    OrderCreateRequest:
      type: object
      required: [poolId, type]
      properties:
        poolId:
          type: integer
          format: int64
        type:
          type: integer
          description: 0-押金 1-车费
        payeeId:
          type: integer
          format: int64
        payeeType:
          type: integer
    OrderVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNo:
          type: string
        userId:
          type: integer
          format: int64
        poolId:
          type: integer
          format: int64
        type:
          type: integer
        amount:
          type: number
        status:
          type: integer
        payeeId:
          type: integer
          format: int64
        payeeType:
          type: integer
        releaseStatus:
          type: integer
        payTime:
          type: string
...
    ScriptCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        type:
          type: string
        difficulty:
          type: integer
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        duration:
          type: integer
        roles:
          type: string
        cover:
          type: string
        priceRef:
          type: number
          format: double
        description:
          type: string
    ScriptVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        difficulty:
          type: integer
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        duration:
          type: integer
        roles:
          type: string
        cover:
          type: string
        priceRef:
          type: number
        description:
          type: string
        status:
          type: integer